refactor: calculate系のモジュールを全て1つのファイルにまとめた
