+++ frontend/src/components/ui/projectList/ProjectListProjectInfo.tsx
@@ -3,8 +3,8 @@ import { occupationList } from 'consts/occupationList'
 import React, { FC } from 'react'
 import styled from 'styled-components'
 import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateMinSizeBasedOnFigmaWidth'
-import { calculateMinSizeBasedOnFigmaHeight } from 'utils/calculateMinSizeBasedOnFigmaHeight'
-import { calculateVhBasedOnFigma } from 'utils/calculateVhBasedOnFigma'
+import { calculateMinSizeBasedOnFigmaHeight } from 'utils/calculateMinSizeBasedOnFigmaWidth'
+import { calculateVhBasedOnFigma } from 'utils/calculateMinSizeBasedOnFigmaWidth'
 import { UserAvatarIcon } from '../avatar/UserAvatarIcon'
 import { UserCount } from '../avatar/UserCount'
